get-server-name: Handle server names and slot counts changing.
This changes the data structure so that the value of the name slots is
the actual name. This means that it's simple (via a data watch) to
change the active set of names. This also makes it so that the script
will bail out properly if the number of slots changes.