Restructure gold info page
Create is_top property for features and partners.
Features with is_top=True will be shown above the
partners list. Partners with is_top=True will be
shown on this page. A link to the remaining partners
is shown after the top partners. The rest of the features
in other_features will be shown below top_partners.