fixup! snoovatars: UI updates.
Setting the box-sizing to 'border-box' for the icon-buttons screwed up the
alignment of the icons.  This sets it only on the orange buttons, which is
really only where it was needed to properly align the two buttons (which were
out of alignment due to one being an <a> element and one being a <button>).