Don't count rg-elves onetime purchases as split transactions
Due to the is_split returning True for RG onetime purchases, the following
transaction was combined with it. The original details of the transaction
(trans_id, pennies, description) are kept, but it's combined with the
subsequent (unrelated) purchase which changes the paying_id and days when
returned.

Not changing the rg-elves type name so that accounting keeps the same
naming scheme.